// src/types.ts
export type UserRole = 'System Admin' | 'Asset Manager' | 'Custodian' | 'Auditor';

export type ConditionCode = 'A1' | 'A2' | 'A3' | 'A4' | 'F1' | 'F2' | 'F3' | 'F4';
export type DepreciationMethod = 'STRAIGHT_LINE' | 'REDUCING_BALANCE' | 'SUM_OF_YEARS';

export interface Asset {
  id: string;
  productId: string;
  name: string;
  category: string;
  acquisitionCost: number;
  acquisitionDate: string;
  netBookValue: number;
  location: string;
  subLocation?: string;
  custodian: string;
  custodianId?: string;
  assignedUser?: string;
  status: 'Active' | 'Disposed' | 'Maintenance' | 'Pending Transfer';
  conditionCode?: ConditionCode;
  imageUrl?: string;
  previousId?: string;
  registrationDate?: string;
  subCategory?: string; // For IT and Office Equipment "Asset Type"
  usefulLife?: number;
  salvageValue?: number;
  method?: DepreciationMethod;
  // Disposal fields
  disposalMode?: 'Sold' | 'Donated' | 'Scrapped' | 'Written Off';
  disposalDate?: string;
  // Improvements logic
  lastImprovementDate?: string;
  improvements?: AssetImprovement[];
  history?: AssetHistoryEvent[];
}

export interface AssetImprovement {
  id: string;
  date: string;
  type: 'Addition' | 'Reduction' | 'Revaluation';
  amount: number;
  description: string;
  newAcquisitionCost: number; // Snapshot of cost after this change
}
// ...
export interface AssetHistoryEvent {
  id: string;
  assetId: string;
  date: string;
  action: string;
  user: string;
  userId: string;
  details: string;
  type: 'Registration' | 'Transfer' | 'Maintenance' | 'Audit' | 'Issue' | 'Update' | 'Condition';
  fromLocation?: string;
  toLocation?: string;
  toCustodian?: string;
}